Enhanceit_Backend/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
bcrypt==4.2.1
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.42.2
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
Cython==3.0.12
dataclasses==0.6
decorator==5.1.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.8
filelock==3.17.0
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
huggingface-hub==0.25.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imantics==0.1.12
imgaug==0.4.0
iopath==0.1.10
ipython==8.32.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
keras-layer-normalization==0.16.0
kiwisolver==1.4.8
labelme2coco==0.2.6
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.20.1
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.1.1
prompt_toolkit==3.0.50
protobuf==5.29.3
pure_eval==0.2.3
pyasn1==0.4.8
pybboxes==0.1.6
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
pyfacer @ git+https://github.com/FacePerceiver/facer.git@ddd35c76ff840174b8a5403ad1c1255e37b8782b
Pygments==2.19.1
PyMatting==1.1.13
pymongo==4.11.1
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
python-slugify==8.0.4
pywin32==308
PyYAML==6.0.2
RealESRGAN @ git+https://github.com/sberbank-ai/Real-ESRGAN.git@362a0316878f41dbdfbb23657b450c3353de5acf
referencing==0.36.2
rembg==2.0.62
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
sahi==0.11.20
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.45.3
stripe==11.6.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminaltables==3.1.10
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2025.2.18
timm==1.0.14
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
xmljson==0.2.1
yacs==0.1.8
